requirements.txt
streamlit
pandas
google-genai
spacy
nltk
vaderSentiment
plotly
pypdf